simu_tools/requirements.txt
Pillow
gym
git+https://github.com/tawnkramer/gym-donkeycar.git
inputs
tensorflow
numba
matplotlib